Implementando outras exceções

+++ backend/src/main/java/br/surb/com/br/dscommerce/services/product/ProductDeleteService.java
@@ -1,7 +1,10 @@
 package br.surb.com.br.dscommerce.services.product;
 
 import br.surb.com.br.dscommerce.repositories.ProductRepository;
+import br.surb.com.br.dscommerce.shared.exeptions.resource.ResourceNotFondExecption;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
@@ -12,8 +15,16 @@ public class ProductDeleteService {
         this.productRepository = productRepository;
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void execute(Long productId){
-        productRepository.deleteById(productId);
+
+        if(!productRepository.existsById(productId)) throw new ResourceNotFondExecption("Entity not found " + productId);
+
+        try {
+            productRepository.deleteById(productId);
+        }catch (DataIntegrityViolationException e){
+            throw new ResourceNotFondExecption("Data integrity violation");
+        }
+
     }
 }
